recipes: lwn_weekly: grab book title from article
Title of weekly editions include release date in heading.
Book title currently appended with book creation date, but
they're not always the same, which could complicate
search if someone has big bunch of magazine issues
in her library.

It's should be easier to follow news if titles on site and
in library are the same.

Signed-off-by: Sergiy Kibrik <sakib@meta.ua>